Crate core_mods

Modules

anchor

Minimal proof of existence registry.

attest

This module allows DIDs to publically attests to arbirary (and arbitrarily large) RDF claimgraphs. These attestations are not stored on-chain; rather, the attester chooses a storage method by specifying an Iri.

blob

Generic immutable single-owner storage.

did

master

Simulates a multisig root account.

revoke

Enums

StateChange

Any state change that needs to be signed is first wrapped in this enum and then its serialized. This is done to prevent make it unambiguous which command was intended as the SCALE codec’s not self describing. Never change the order of variants in this enum

Type Definitions

BlockNumber